WIP: Replace REBUILD_ACTION with more tasks
Actually, not sure I want to carry on with this.
If you're testing machine by machine, then it makes sense, but if you
have multiple machines, not so much.
Wanted to make an hybrid aproach at the end, but then it just makes
things more complex without a really good reason.